@@ -16,172 +16,110 @@
     "format": "prettier --write './**/*.{ts,tsx,js,jsx,json}'",
     "format-and-lint": "npm run format && npm run lint:style && npm run lint:ts:fix",
     "start": "BROWSER=none REACT_APP_MODE=live node scripts/start.js",
-    "start:demo": "BROWSER=none REACT_APP_MODE=dev node scripts/start.js",
+    "start:demo": "VITE_APP_MODE=dev vite",
     "test": "node scripts/test.js",
     "test:lib": "jest './libs/*' --colors --coverage --watchAll",
     "test:schema": "vitest './src/schema/box.schema.spec.ts'"
   },
   "dependencies": {
-    "@babel/core": "^7.21.4",
-    "@babel/plugin-transform-typescript": "^7.21.3",
-    "@babel/preset-env": "^7.21.4",
-    "@babel/preset-react": "^7.18.6",
-    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.10",
-    "@svgr/webpack": "^5.5.0",
-    "@typescript-eslint/parser": "^5.59.1",
-    "@zodios/core": "^10.9.0",
-    "axios": "^1.3.6",
+    "@babel/core": "^7.23.9",
+    "@babel/plugin-transform-typescript": "^7.23.6",
+    "@babel/preset-env": "^7.23.9",
+    "@babel/preset-react": "^7.23.3",
+    "@typescript-eslint/parser": "^7.0.2",
+    "@vitejs/plugin-react": "^4.2.1",
+    "@zodios/core": "^10.9.6",
+    "axios": "^1.6.7",
     "babel-core": "^6.26.3",
     "babel-eslint": "^10.1.0",
-    "babel-jest": "^27.5.1",
-    "babel-loader": "^8.3.0",
-    "babel-plugin-macros": "^2.8.0",
+    "babel-jest": "^29.7.0",
+    "babel-loader": "^9.1.3",
+    "babel-plugin-macros": "^3.1.0",
     "babel-plugin-named-asset-import": "^0.3.8",
     "babel-plugin-transform-export-extensions": "^6.22.0",
     "babel-polyfill": "^6.26.0",
     "babel-preset-env": "^1.7.0",
     "babel-preset-react-app": "^10.0.1",
-    "bfj": "^7.0.2",
-    "browserslist": "^4.21.5",
-    "camelcase": "^6.3.0",
-    "case-sensitive-paths-webpack-plugin": "^2.4.0",
-    "css-loader": "^6.7.3",
-    "css-minimizer-webpack-plugin": "^3.4.1",
-    "date-fns": "^2.29.3",
-    "dotenv": "^10.0.0",
-    "dotenv-expand": "^5.1.0",
-    "eslint": "^8.39.0",
+    "browserslist": "^4.23.0",
+    "camelcase": "^8.0.0",
+    "css-loader": "^6.10.0",
+    "date-fns": "^3.3.1",
+    "dotenv": "^16.4.5",
+    "dotenv-expand": "^11.0.6",
+    "eslint": "^8.57.0",
     "eslint-config-react-app": "^7.0.1",
-    "eslint-webpack-plugin": "^3.2.0",
     "fetch-mock": "^9.11.0",
     "file-loader": "^6.2.0",
-    "fs-extra": "^10.1.0",
-    "html-webpack-plugin": "^5.5.1",
+    "fs-extra": "^11.2.0",
     "identity-obj-proxy": "^3.0.0",
-    "jest": "^27.5.1",
-    "jest-resolve": "^27.5.1",
-    "jest-watch-typeahead": "^1.1.0",
+    "jest": "^29.7.0",
+    "jest-resolve": "^29.7.0",
     "lodash": "^4.17.21",
-    "mini-css-extract-plugin": "^2.7.5",
-    "miragejs": "^0.1.47",
-    "modern-normalize": "^1.1.0",
+    "mini-css-extract-plugin": "^2.8.0",
+    "modern-normalize": "^2.0.0",
     "moxios": "^0.4.0",
-    "node-fetch": "^2.6.9",
+    "msw": "^2.2.1",
+    "node-fetch": "^3.3.2",
     "os-browserify": "^0.3.0",
     "path-browserify": "^1.0.1",
-    "postcss": "^8.4.23",
+    "postcss": "^8.4.35",
     "postcss-flexbugs-fixes": "^5.0.2",
-    "postcss-loader": "^6.2.1",
+    "postcss-loader": "^8.1.0",
     "postcss-normalize": "^10.0.1",
-    "postcss-preset-env": "^7.8.3",
-    "prettier": "^2.8.8",
-    "process": "^0.11.10",
+    "postcss-preset-env": "^9.4.0",
+    "prettier": "^3.2.5",
     "prompts": "^2.4.2",
-    "react": "^16.14.0",
+    "react": "^18.2.0",
     "react-app-polyfill": "^3.0.0",
     "react-dev-utils": "^12.0.1",
-    "react-dom": "^16.14.0",
-    "react-refresh": "^0.11.0",
-    "react-transition-group": "^1.2.1",
-    "regenerator-runtime": "^0.13.11",
-    "resolve": "^1.22.2",
-    "resolve-url-loader": "^4.0.0",
-    "sass-loader": "^12.6.0",
-    "semver": "^7.5.0",
-    "source-map-loader": "^3.0.2",
-    "style-loader": "^3.3.2",
-    "styled-components": "^5.3.10",
-    "tailwindcss": "^3.3.1",
-    "terser-webpack-plugin": "^5.3.7",
-    "ts-node": "^10.9.1",
-    "typescript": "^5.0.4",
-    "uid": "^1.0.0",
-    "vite": "^4.3.3",
-    "vitest": "^0.30.1",
-    "webpack": "^5.80.0",
-    "webpack-dev-server": "^4.13.3",
-    "webpack-manifest-plugin": "^4.1.1",
-    "workbox-webpack-plugin": "^6.5.4",
-    "zod": "^3.21.4"
+    "react-dom": "^18.2.0",
+    "react-refresh": "^0.14.0",
+    "react-transition-group": "1.x",
+    "regenerator-runtime": "^0.14.1",
+    "resolve": "^1.22.8",
+    "resolve-url-loader": "^5.0.0",
+    "sass-loader": "^14.1.1",
+    "semver": "^7.6.0",
+    "source-map-loader": "^5.0.0",
+    "style-loader": "^3.3.4",
+    "styled-components": "^6.1.8",
+    "tailwindcss": "^3.4.1",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.3.3",
+    "uid": "^2.0.2",
+    "vite": "^5.1.4",
+    "vite-plugin-svgr": "^4.2.0",
+    "vite-tsconfig-paths": "^4.3.1",
+    "vitest": "^1.3.1",
+    "zod": "^3.22.4"
   },
   "devDependencies": {
-    "@types/jest": "^25.2.3",
-    "@types/lodash": "^4.14.194",
-    "@types/node": "^13.13.52",
-    "@types/prop-types": "^15.7.5",
-    "@types/react": "^16.14.40",
-    "@types/react-dom": "^16.9.18",
-    "@types/react-transition-group": "^4.4.5",
-    "@types/styled-components": "^5.1.26",
-    "@typescript-eslint/eslint-plugin": "^5.59.1",
-    "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.7",
-    "enzyme-to-json": "^3.6.2",
+    "@types/jest": "^29.5.12",
+    "@types/lodash": "^4.14.202",
+    "@types/node": "^20.11.20",
+    "@types/react": "^18.2.58",
+    "@types/react-dom": "^18.2.19",
+    "@types/react-transition-group": "^4.4.10",
+    "@types/styled-components": "^5.1.34",
+    "@typescript-eslint/eslint-plugin": "^7.0.2",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-plugin-flowtype": "^8.0.3",
-    "eslint-plugin-import": "^2.27.5",
-    "eslint-plugin-jsx-a11y": "^6.7.1",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-jsx-a11y": "^6.8.0",
     "eslint-plugin-lodash": "^7.4.0",
-    "eslint-plugin-react": "^7.32.2",
+    "eslint-plugin-react": "^7.33.2",
     "eslint-plugin-react-hooks": "^4.6.0",
-    "jest-styled-components": "^7.1.1",
-    "react-test-renderer": "^16.14.0",
-    "stylelint": "^13.13.1",
+    "jest-styled-components": "^7.2.0",
+    "stylelint": "^16.2.1",
     "stylelint-config-rational-order": "^0.1.2",
-    "stylelint-config-standard": "^20.0.0",
+    "stylelint-config-standard": "^36.0.0",
     "stylelint-config-styled-components": "^0.1.1",
-    "stylelint-order": "^4.1.0",
+    "stylelint-order": "^6.0.4",
     "stylelint-processor-styled-components": "^1.10.0"
   },
-  "jest": {
-    "roots": [
-      "<rootDir>/src"
-    ],
-    "collectCoverageFrom": [
-      "src/**/*.{js,jsx,ts,tsx}",
-      "!src/**/*.d.ts"
-    ],
-    "setupFiles": [
-      "react-app-polyfill/jsdom"
-    ],
-    "setupFilesAfterEnv": [
-      "<rootDir>/src/setupTests.js"
-    ],
-    "testMatch": [
-      "<rootDir>/src/**/__tests__/**/*.{js,jsx,ts,tsx}",
-      "<rootDir>/src/**/*.{spec,test}.{js,jsx,ts,tsx}"
-    ],
-    "testEnvironment": "jsdom",
-    "transform": {
-      "^.+\\.(js|jsx|mjs|cjs|ts|tsx)$": "<rootDir>/config/jest/babelTransform.js",
-      "^.+\\.css$": "<rootDir>/config/jest/cssTransform.js",
-      "^(?!.*\\.(js|jsx|mjs|cjs|ts|tsx|css|json)$)": "<rootDir>/config/jest/fileTransform.js"
-    },
-    "transformIgnorePatterns": [
-      "[/\\\\]node_modules[/\\\\].+\\.(js|jsx|mjs|cjs|ts|tsx)$",
-      "^.+\\.module\\.(css|sass|scss)$"
-    ],
-    "modulePaths": [],
-    "moduleNameMapper": {
-      "^react-native$": "react-native-web",
-      "^.+\\.module\\.(css|sass|scss)$": "identity-obj-proxy"
-    },
-    "moduleFileExtensions": [
-      "web.js",
-      "js",
-      "web.ts",
-      "ts",
-      "web.tsx",
-      "tsx",
-      "json",
-      "web.jsx",
-      "jsx",
-      "node"
-    ],
-    "watchPlugins": [
-      "jest-watch-typeahead/filename",
-      "jest-watch-typeahead/testname"
-    ],
-    "resetMocks": true
+  "msw": {
+    "workerDirectory": [
+      "public"
+    ]
   }
 }
