@@ -3,6 +3,14 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 
 import './index.css';
+import 'modern-normalize';
+
 import { App } from './components/App';
 
-ReactDOM.render(<App />, document.getElementById('root'));
+const root = ReactDOM.createRoot(document.getElementById('root'));
+
+root.render(
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>,
+);
